Status indicator for enabled
